// ui/src/generated/models/media.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Photos NG API
 * API for managing albums and media in Photos NG application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ExifHeader } from './exif-header';

/**
 * 
 * @export
 * @interface Media
 */
export interface Media {
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    'href': string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    'albumHref': string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    'capturedAt': string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    'type': string;
    /**
     * full path of the media file on the disk
     * @type {string}
     * @memberof Media
     */
    'filename': string;
    /**
     * href to thumbnail
     * @type {string}
     * @memberof Media
     */
    'thumbnail': string;
    /**
     * href of the endpoint serving the content of the media
     * @type {string}
     * @memberof Media
     */
    'content': string;
    /**
     * 
     * @type {Array<ExifHeader>}
     * @memberof Media
     */
    'exif': Array<ExifHeader>;
}